+++ index.html
@@ -1,8 +1,10 @@
 <!doctype html>
 <html>
   <head>
+    <!-- build:css /css/styles.css -->
     <link rel="stylesheet" type="text/css" href="jspm_packages/github/twbs/bootstrap@3.3.4/css/bootstrap.min.css">
     <link rel="stylesheet" type="text/css" href="css/main.css">
+    <!-- endbuild -->
   </head>
   <body>
     <div class="container" ng-controller="MainCtrl as ctrl">
@@ -15,10 +17,27 @@
       </div>
       <flickr-gallery data="ctrl.service.images" />
     </div>
-    <script src="jspm_packages/system.js"></script>
-    <script src="config.js"></script>
+    <!-- @if NODE_ENV='development' -->
+    <script src="/jspm_packages/system.js"></script>
+    <script src="/config.js"></script>
+    <script>
+      System.config({
+        "paths": {
+          "*": "/*.js",
+          "github:*": "/jspm_packages/github/*.js",
+          "npm:*": "/jspm_packages/npm/*.js",
+          "angular-es6/*": "/lib/*.js"
+        }
+      });
+    </script>
     <script>
       System.import('./dist/app').catch(console.error.bind(console));
     </script>
+    <!-- @endif -->
+    <!-- @if NODE_ENV='production' -->
+    <script src="/system.js"></script>
+    <script src="/config.js"></script>
+    <script src="/app.js"></script>
+    <!-- @endif -->
   </body>
 </html>
