+++ package.json
@@ -12,12 +12,16 @@
   "main": "index.js",
   "devDependencies": {
     "browser-sync": "^1.8.1",
+    "del": "^1.2.0",
     "gulp": "^3.8.10",
     "gulp-babel": "^5.1.0",
     "gulp-changed": "^1.1.0",
+    "gulp-exec": "^2.1.1",
     "gulp-jshint": "^1.9.0",
     "gulp-plumber": "^0.6.6",
+    "gulp-preprocess": "^1.2.0",
     "gulp-sourcemaps": "^1.5.2",
+    "gulp-useref": "^1.2.0",
     "jasmine-core": "^2.1.3",
     "jshint-stylish": "^1.0.0",
     "karma": "^0.12.28",
@@ -36,8 +40,9 @@
       "bootstrap": "github:github:twbs/bootstrap@^3.3.1"
     },
     "devDependencies": {
-      "traceur": "github:jmcriffey/bower-traceur@0.0.88",
-      "traceur-runtime": "github:jmcriffey/bower-traceur-runtime@0.0.88"
+      "babel": "npm:babel-core@^5.5.6",
+      "babel-runtime": "npm:babel-runtime@^5.5.6",
+      "core-js": "npm:core-js@^0.9.4"
     }
   }
 }
